requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
torch==2.2.0+cpu --extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.17.0+cpu --extra-index-url https://download.pytorch.org/whl/cpu
pillow==10.2.0
httpx==0.26.0
streamlit==1.31.0
pandas==2.2.0
filelock==3.13.1